Fix the Travis build
Docker controls the environment. By specifying Node, Travis
automatically install npm modules, which then causes issues if the
Docker version of Node is different. So let's juse use C as the Travis
language since Travis doesn't treat Docker as a first class citizen (not
specifying a language defaults to a Ruby environment).

Also downgrade Node back to v6. I forgot that v7/v8 causes weird issues
with some native modules, and it's not worth the time to keep trying to
fix it.