Rollback autotrack to fix minification
In autotrack v2, the published package is in the form of ES6 modules,
and this is causing an issue with Webpack minification.

See https://github.com/googleanalytics/autotrack/issues/137

I tried the proposed solution of excluding it from Babel, but it
didn't work on the first try, and I don't want to spend any more
time on it. We only use one autotrack plugin anyway.