Limit results to avoid a NewsAPI error
We're getting this error: "You have requested too many results.
Developer accounts are limited to a max of 100 results. You are trying
to request results 100 to 200. Please upgrade to a paid plan if you need
more results."

For now, only use the first 100 results, and don't paginate. To get the
full results again, we can try to make multiple queries, each with a
subset of the sources.